Sampling points along each ruelle (3 per ruelle)
This is a .shp file containing the exact points to be sampled in the ruelles. There are 3 per sites per ruelle, with the exception of a few ruelles that were oddly shaped and I felt it was best to have more than 3 to account for the unusual shape. If we want, we can delete the extra sites later if we do not want to use them.